+++ src/tests/unit/service/test_service.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import sessionmaker
 import model.models as models
 import schema.schemas as schemas
 # from service.services import get_dummy, get_dummies, create_dummy, delete_dummy, update_dummy
-import service.dummy as services
+import service.services as services
 
 
 @pytest.fixture(scope="function")
